Update README.md
The exploit allows you to convert EXE to .PDF file its coded 100% from scratch and used by private methods to assure a great stability and long lasting FUD time. You are able to attach it to all email providers and now a days everyone uses Adobe based Reader or PDF Reader so it gives a huge chance of success.